[Refactor] 마감임박 쿠폰 조회 패키지 market -> coupon 변경 - #101
