+++ src/main/java/com/appcenter/marketplace/domain/coupon/service/CouponService.java
@@ -1,6 +1,8 @@
 package com.appcenter.marketplace.domain.coupon.service;
+
 import com.appcenter.marketplace.domain.coupon.dto.res.CouponMemberRes;
 import com.appcenter.marketplace.domain.coupon.dto.res.CouponPageRes;
+import com.appcenter.marketplace.domain.coupon.dto.res.TopClosingCouponRes;
 
 import java.util.List;
 
@@ -8,4 +10,5 @@ public interface CouponService {
 
    CouponPageRes<CouponMemberRes> getCouponList(Long marketId, Long couponId, Integer size);
 
+   List<TopClosingCouponRes> getTopClosingCoupons(Integer size);
 }
