+++ src/main/java/com/appcenter/marketplace/domain/market/repository/MarketRepositoryCustom.java
@@ -1,6 +1,9 @@
 package com.appcenter.marketplace.domain.market.repository;
 
-import com.appcenter.marketplace.domain.market.dto.res.*;
+import com.appcenter.marketplace.domain.market.dto.res.LatestCouponRes;
+import com.appcenter.marketplace.domain.market.dto.res.MarketDetailsRes;
+import com.appcenter.marketplace.domain.market.dto.res.MarketRes;
+import com.appcenter.marketplace.domain.market.dto.res.TopLatestCouponRes;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -26,7 +29,6 @@ public interface MarketRepositoryCustom {
     List<TopLatestCouponRes> findTopLatestCoupons(Long memberId, Integer size);
 
     List<LatestCouponRes> findLatestCouponList(Long memberId, LocalDateTime createdAt, Long couponId, Integer size);
-
-    List<TopClosingCouponRes> findTopClosingCoupons(Integer size);
+    
 }
