+++ src/main/java/com/appcenter/marketplace/domain/market/repository/MarketRepositoryCustomImpl.java
@@ -329,44 +329,6 @@ public class MarketRepositoryCustomImpl implements MarketRepositoryCustom{
                 .fetch();
     }
 
-    // 마감 임박 쿠폰 TOP 조회
-    @Override
-    public List<TopClosingCouponRes> findTopClosingCoupons(Integer size) {
-
-        QCoupon subCoupon = new QCoupon("subCoupon");
-
-        // 서브쿼리: 각 market_id 그룹별 가장 가까운 deadLine을 구함
-        JPQLQuery<Tuple> subQuery = JPAExpressions
-                .select(subCoupon.market.id, subCoupon.deadLine.min())
-                .from(subCoupon)
-                .innerJoin(subCoupon.market, market)
-                .where(subCoupon.isDeleted.eq(false)
-                        .and(subCoupon.isHidden.eq(false))
-                        .and(subCoupon.stock.gt(0))
-                        .and(subCoupon.deadLine.after(LocalDateTime.now())))
-                .groupBy(subCoupon.market.id)
-                .limit(1);
-
-        return jpaQueryFactory.select(new QTopClosingCouponRes(
-                        market.id,
-                        coupon.id,
-                        market.name,
-                        coupon.name,
-                        coupon.deadLine,
-                        market.thumbnail))
-                .from(coupon)
-                .innerJoin(coupon.market, market)
-                .where(Expressions.list(coupon.market.id, coupon.deadLine).in(subQuery)
-                        .and(coupon.isDeleted.eq(false))
-                        .and(coupon.isHidden.eq(false))
-                        .and(coupon.stock.gt(0))
-                        .and(coupon.deadLine.after(LocalDateTime.now())))
-                .orderBy(coupon.deadLine.asc(), coupon.id.desc())
-                .limit(size)
-                .fetch();
-    }
-
-
     // BooleanExpression을 반환 시 where의 첫 조건에서 null 예외가 뜰 수 있다.
     // lt= less than = <(~보다 작은)
     private BooleanBuilder ltMarketId(Long marketId){
